requirements.txt
agno>=0.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
